星流平台(aicp)：GPU虚拟化的开发任务及节点列表OpenAPi变动;

+++ src/service/Aicp/v20251212/index.js
@@ -37,7 +37,7 @@ module.exports = class Client extends BaseClient {
         Framework: "String",
         AccessType: "String",
         SelfHealing: "Boolean",
-        MaxRuntimeHour: "Int",
+        MaxRuntimeHour: "Long",
         JobRunOnCPU: "Boolean",
         SupportTensorboard: "Boolean",
         StorageConfigs: "Array",
